nutrisync-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pydantic>=2.7.0
email-validator>=2.1.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
requests>=2.32.0
redis>=5.0.4
pandas>=2.2.2
numpy==1.26.4
scikit-learn>=1.5.0
scikit-surprise>=1.1.4
Pillow>=10.3.0
torch==2.2.2
torchvision==0.17.2
bcrypt==4.0.1
